requirements.txt
Django
PIL
South
django-taggit
sorl-thumbnail
psycopg2
python-memcached
fabric
ipython